Bugreport treatAsAdminAnyway override in test
We now temporarily treat HSU as Admin in select places, including for bugreports. This can cause tests to fail, since they may implicitly run in the system user, so we override this so that tests have consistent behaviour (where the Admin status can be chosen explicitly for the test).


Bug: 457559134
Bug: 461646697
Bug: 467128801
Flag: android.multiuser.hsu_not_admin
Test: atest BugreportManagerTest#testRemoteBugreport
Test: atest BugreportManagerServiceImplTest
Change-Id: I5b3a101f826555e7ebb3b18061cc10fbd51afc1e