Enable fluid resizing for listed apps
Introduce the feature flag to enable the per-app override for the fluid
resizing. Besides, a new app compat flag is introduced to provide a way
to let the device to override the behavior for the apps as needed.

Updated the implementation to apply the corresponding TaskPositioner
based on the flag values.

Bug: 460405642
Test: Override the flags and test it works for app
Flag: com.android.window.flags.enable_fluid_resizing_for_listed_apps

Change-Id: Ia7fdaea5800b16981d2fa229bf368f5e3a3cd47d