Launch activities with CC owner permissions
Currently the system context is used to startActivity (even though
the activity is started as the session owner user). So the start
privileges are taken from the system process. Instead, pass the
IApplicationThread object from the app process directly into AMS
startActivityAsUser so there's no transfer of privleges.

Caching IApplicationThread in ComputerControlSessionImpl is safe
because the session cannot outlive the app process (it's bound to
the session callback, which comes from the app process)

Bug: 461765546
Fix: 461765546
Test: atest & manual
Flag: android.companion.virtualdevice.flags.computer_control_access

Change-Id: I4050d0202c0db9c7a58f746119cffd9f52016bf2