Optimize drawing, remove NativePaint and harden withoutReadObservation
1. Use drawWithCache to optimize drawing
2. Use TextMeasurer instead of NativePaint and NativeCanvas
3. Due to moving things into `onDrawWithContent` the subscription
within the draw phase to the debugConfig states has changed. We need
to make sure that we are not too broad with the withoutReadObservation.
The key is passed as value anyways so there is nothing to protect
against. For the stl/state we want to make sure that reads are always
unsubscribed. `WithoutReadObservation<T>` helps enforcing this in a
concise way.

Bug: 463679218
Test: Manually tested the app, all configurations work
Flag: NONE - debugger App
Change-Id: I0f81e6eadd79ee1012c7da20d2ccf403673304e4