Add an HSUM check to MasterClearReceiverTests
In MasterClearReceiver, a check is done to see if the calling user is
not the system user and to check if the system is in headless system
user mode (HSUM). However, in MasterClearReceiverTest, when UserManager
is mocked, the call to UserManager.isHeadlessSystemUserMode() is only
mocked in one specific test. Because mockito returns false by default
for a boolean method, this means that when these tests are run on
automotive devices, MasterClearReceiver believes that they are not
running HSUM even though they are, which causes some of the tests to
fail.

- Update the tests to use ExtendedMockitoRule.
- Update expectHeadlessSystemUserMode to take a boolean and set the
  return value of isHeadlessSystemUserMode to that boolean.
- Update testNonSystemUser to mock isHeadlessSystemUserMode as false.
- Add a mock for ActivityManager.getCurrentUser() to return
  UserHandle.USER_SYSTEM when testWipeNonSystemUser runs so that the
  system doesn't attempt to switch to user 0, which would fail on HSUM
  devices.

Test: atest MasterClearReceiverTest
Flag: EXEMPT TEST_ONLY
Fixes: 467351079
Change-Id: I52bc0de29b5b751a74acc5d67d21d4c38d6c50a5