Accessibility resizers should stay on after resizing
In this CL, we change `CommunalHub.kt` to keep the ResizeableItemFrameViewModel attached to the widget after resizing, as we need to preserve the state following whether we have the accessibility resizer on or not. Previously, the ViewModel was recreated after resizing.

Flag: com.android.systemui.communal_edit_mode_accessibility_resize
Bug: 431683508
Test: flashed Blazer & TangorPro devices to check that the accessibility resizer stays on after resizing.
Change-Id: I1cb8c7d23a035afc5b9559d884527a786baa1163