Rewrite migration for internet tile
Due to change in requirements the migration needs to take into account
tile size, in order to determine if it migrates just to wifi, or wifi +
cell.

In order to achieve this, we perform the migration when defining the
flow of tiles in UserTileSpecRepository. That way, all downstream
consumers will see the correct set of tiles. As an additional benefit,
this is applied to all changes (so we get restore for free). Note that
the set of large tiles is manipulated as part of the migration.

For details on the specific scenarios of the migration, see
InternetTileMigration.kt.

Test: manual, enable and disable flag
Test: atest com.android.systemui.qs
Flag: com.android.systemui.qs_split_internet_tile
Fixes: 466154446
Change-Id: I609f61581ad2b2b8f1c014b1c8ad0f85ca749203