Device Idle: avoid exiting Idle when on-body if screen is not on
Before this change, there is a logic to force the Device Idle state to
be ACTIVE whenever the device becomes on-body. Given that the on-body
logic is only used by Wear, this basically means that, whenever the user
wears the watch, Idle mode is exited. However, this has an issue,
because the ACTIVE state remains until there is reason to exit ACTIVE
state. So, if the user simply wears the watch and goes about their day
without interacting with their watch, their device never enters Idle
mode. With this fix, we exit Idle mode when the watch is worn only if
the screen is on. That way, we avoid indefinite Idle mode states,
because, if the screen is on, the device will eventually start
transitioning into Idle mode once the screen times out.

Bug: 468890229
Flag: com.android.server.power.feature.flags.interactive_doze_experience
Test: atest DeviceIdleControllerTest
Change-Id: I857cffe4198b55da48fce7ccf4b4d61377e22b25