Track BurnInMovement as hydrated state
This simplifies rememberBurnIn and burnInAware to better seperate
concerns between the two methods. rememberBurnIn has been replaced
with trackBurnInParameters which tracks relevant view parameters
which are used as input for burn-in. burnInAware still produces a
Modifier responsible for correctly rendering an element with burn-in
but is much simpler as it no longer transforms the state in any way.

Bug: 468487364
Flag: com.android.systemui.scene_container
Test: Manually checked for a regression in b/371211911
Change-Id: Ifb5b8041d9c35a1043e969fb090e455111eb253f