Add initiate_bugreport_as_non_admin feature flag
This CL introduces a new Aconfig flag, `initiate_bugreport_as_non_admin`, to control the ability of non-admin users to initiate bug reports.

This is the first step in a larger effort to allow designated apps to generate bug reports on multi-user devices, even when the current user is not an administrator.

Flag: android.os.initiate_bugreport_as_non_admin
Bug: 450735514
Test: None
Change-Id: I346d12b25fccf63d98e193bb84950c6d2399cce5