Support Telecom requesting focus using requestAudioFocus API.
Telecom used `requestAudioFocusForCall` in the past, which is a hidden API.
Ultimately we DO NOT want to implement an `AudioPolicy` and to control
all focus on the device.  In the mid term the API contract changes being
made between Audio and Telecom will remove the necessity for this change
but in the short term we need to build Telecom as an APEX and this dep
needs to be cleaned up.

I'm not super happy about the hard coded package name in AudioManager, but
that check fails fast.

In `AudioService`, changed `requestAudioFocus` to allow Telecom as a caller
to go into the `MODIFY_PHONE_STATE` permission check which comes next.
This enables us to request focus without providing an `AudioPolicy`.

Bug: 469224275
Test: Ran Telecom audio focus CTS tests.
Test: Manual audio focus verification.
Flag: EXEMPT PURE_REFACTOR
Change-Id: I1b0b7b3e52c0ebae7d9435aa9360ced6ac4324b3