Allowlist INTERACT_ACROSS_USERS permission for contactspicker
This permission is needed to query contacts across profiles

Starting with Android 8.0 (and strictly enforced in Android 9+), the system checks all apps in the /system/priv-app/ directory during boot. If a privileged app requests a permission with the protectionLevel="signature|privileged" (like INTERACT_ACROSS_USERS) but that permission is not explicitly listed in a privapp-permissions-*.xml file, the system throws a fatal exception to prevent potential security breaches. This fatal exception crashes the SystemServer or PackageManager, causing the device to reboot endlessly.

Bug: 469380160
Flag: android.content.flags.enable_system_contacts_picker
Test: N/A
Change-Id: I44168c75bf2420963746f5c44ce54ae6d746076e