Add light mute logic to LightsService.
Some system services like the CameraService have the need to mute the
lights for short periods of time while they take a picture or video,
so that the lights don't disturb the scene.

This change adds the logic in the LightsService to mute the lights
temporarily.

Some lights are considered critical, such as keyboard or backlight, and
these will not be shutdown.

Bug: 456835809
Fixed: 456835809
Flag: com.android.server.lights.feature.flags.enable_light_animations
Test: atest FrameworksServicesTests:LightsServiceTest
Test: presubmit.
Change-Id: Ic54a20bf3b7c824b6dd2891fcbe4652e5f4e6618