Reland "Apply rotation transform if the orientation source is not top"
For example, when display is in portrait, move a task to front:
 B: Translucent activity orientation=unspecified
 A: Opaque activity orientation=landscape
B does not provide orientation, so the orientation source is A which
is landscape. Then when resuming B, it should be applied with
landscape transform. And when making A visible, it should share the
same transform from B, because the orientation of orientation source
should be the same as the orientation of the launching fixed
rotation app. And if the orientation of non-top visible activity is
unspecified, the activity either stay in current display orientation
or use the same orientation as the launching fixed rotation app which
will affect the display orientation.

Note that updateOrientation can be called when
 top INITIALIZING landscape visibleRequested=false
     FINISHING landscape visibleRequested=false
     PAUSING portrait visibleRequested=true
The orientation source is the portrait activity.
But the top landscape activity should not apply portrait transform.
So the condition of getTopActivityFollowingBehindOrientation avoids
the case.

Fix: 464307133
Flag: EXEMPT BUGFIX
Test: atest DisplayContentTests#testUnspecifiedOrientationTranslucentTop
      atest PreferenceActivityFlowLandscapeTest

Change-Id: I63c40134ceb50249e954eefb9d010f56473d4bfc