Refactor createMergedTxts(...)
Previously, it took `stubsTypeSuffix` which differed between exportable
and non-exportable files. This replaces that with `exportableStubsType`
and computes the `stubsTypeSuffix` based on that. That is in
preparation for a follow-up change that will use the
`exportableStubsType` to construct the tag so it can access
exportable signature files from the modules.

Flag: EXEMPT HOST_MODE
Bug: 425925746
Test: m checkapi
Change-Id: Ie914bb4493415f4e50f133e075e0610c63f16f97