Remove @Composable default param which is crashlooping studio builds
Apparently there is some configuration/dependency problem with sysui-
studio builds that produce a crash (see b/469745301). The odd thing is
that the function returns a non-null value but the caller receives a
null value. So there has to be some issue in byte code.

It seems the issue is very rare as the following conditions need to be
met:
1. @Composable interface function
2. With default parameters
3. The default param needs to be overwritten by caller
4. The function needs to return something

In this case the easy fix is to just remove the default param. I believe
a closer investigation is not warranted at this point since the above
conditions are rare enough and the issue will probably be resolved when
we update some dependency in sysui studio (soong builds are not affected
since it seems we use other prebuilts there).

Bug: 469745301
Test: Crash loop gone when building with sysui studio
Test: ElementTest#previewAlphaTransformation is now working in sysui
studio (previously has been broken for same reason)
Flag: EXEMPT Bugfix

Change-Id: Ib9dfe7d8521b5b22bdf632142c833fe1d4a32a82