Update deviceEntryParentViewAlpha onCancel in KTF transitions
deviceEntryParentViewAlpha should always be updated onCancelled
to ensure the correct starting state for the next transition,
because the next transition will always start from the last `to` state.

For example:
1. [start] Lockscreen => Primary Bouncer
2. [cancelled] Lockscreen => Primary Bouncer
3. [start] Primary Bouncer => Occluded

In the above example, after [2] the deviceEntryParentViewAlpha should
be set to 0 when the transition gets cancelled. [3] assumes that the
deviceEntryParentViewAlpha was never showing on the Primary Bouncer,
so there's no alpha update when transitioning to Occluded.

Fixes: 466215375
Flag: EXEMPT BUGFIX
Test: atest LockscreenToPrimaryBouncerTransitionViewModelTest
Change-Id: Ib69c74861e7516aae89dc75adfeb4b6d60b4c6b6