Change bubble unsupported behavior for invalid apps.
This updates the UX to collapse the stack when folded with an invalid app, consistent with b/463258978#comment4. If there is a tap on that bubble on fold, the app moves out of a bubble to the front and goes into fullscreen with an error toast. If there is no interaction on fold, the app bubble will remain when unfolded.

Also add isTaskValidToBubbleOnSmallScreen to skip
BarToFloatingTransition if the task is not valid to bubble on small
screen.

Demo: https://screencast.googleplex.com/cast/NDY4NjA0MjAyNzI2MTk1MnxjMTA4MzVkMi0wNA
Bug: 468934749
Bug: 436326529
Test: atest WMShellMultivalentTestsOnDevice:BubbleBarExpandedViewTest
Test: atest WMShellUnitTests:DefaultBubblePolicyHelperTest
Flag: com.android.wm.shell.enable_create_any_bubble
Change-Id: If7798f1dc804cb95857322eefa5c25e8f3d104c9