[flexiglass] Add end state transitions for keyguard transitions
This change introduces two new "end state" transition view models:
 - `ToAodEndStateTransitionViewModel`
 - `ToDozingEndStateTransitionViewModel`
 - `ToLockscreenEndStateTransitionViewModel`

These view models act as a fallback to ensure that all views are in
their correct final state when the AOD, Dozing or Lockscreen states
are settled.
This helps to prevent UI inconsistencies and ensures a stable state.

Bug: 462086517
Test: manual test - launch an activity from the lockscreen (such as
by clicking a notification and unlocking the device) and then
immediately press the power button to lock the device. Without this CL,
sometimes lockscreen ends up in bad state (blankscreen or miss
something); With this CL, lockscreen is good.
Flag: com.android.systemui.scene_container

Change-Id: I9dd1dddab41f4b2c12ab2e27bf9676e78052b2ce