Notify IME Switcher Menu of IME and Subtype change
This notifies the IME Switcher Menu when the selected IME and subtype of
a given user changes. This allows the menu UI to be updated in real
time, if the change happens while the menu is shown (e.g. through the
CTRL + Space keyboard shortcut). When the new one is selected, the UI
also scrolls to the new selection.

Flag: android.view.inputmethod.ime_switcher_menu_systemui
Bug: 460776726
Test: show the Menu; adb shell input keycombination CTRL_LEFT SPACE
  atest InputMethodManagerTest#testInputMethodPickerShownItems
  InputMethodManagerTest#testInputMethodPickerSwitchIme
  InputMethodServiceTest#testImeSwitchButtonLongClick
  InputMethodServiceTest#testImeSwitcherMenu_openLanguageSettings
  InputMethodServiceTest#testImeSwitcherMenu_noLanguageSettingsWhenScreenLocked
  InputMethodServiceTest#testImeSwitcherMenu_noLanguageSettingsWhenDeviceNotProvisioned
Change-Id: Id5872e5f16d2fe9a7fcecb84b1e8365a12724859