Exempt HSU bugreport only for REMOTE
The HSU is not an Admin, and that means that it cannot take bugreports
anymore. We therefore exempted the SYSTEM [1]. But the only known problem
is with REMOTE bugreports, so let's un-relax the exemption to be
specifically for that case. We guard this under a separate flag in case
this proves problematic.

[1] See I5b3a101f826555e7ebb3b18061cc10fbd51afc1e.

Bug: 457559134
Flag: android.multiuser.hsu_not_admin_for_bugreports

Test: atest BugreportManagerTest#testRemoteBugreport
Test: atest BugreportManagerServiceImplTest
Test: atest CtsDevicePolicyTestCases:android.devicepolicy.cts.BugReportTest

Change-Id: I1a73e0cbcdcbf279506aa7c4e46de41d22128b6f