Make StatusBarHeadsUpChangeListener display aware
This makes sure that the calls to forcing the status bar visible when
a HUN is pinned, is done on the correct display, i.e. the display where
the shade is showing.

It also handles moving of the shade between displays while a HUN is
already pinned.

This also removes the caching of the last display id in
ShadeControllerImpl, as it is not needed anymore with this change.

Test: StatusBarHeadsUpChangeListenerTest.kt
Test: Trigger a HUN and check the status bar is visible
Test: Trigger a HUN on an external display and check the status bar and
      HUN are visible on the external display
Test: Trigger a HUN and move the shade to a different display
Fixes: 432399588
Flag: com.android.systemui.display_aware_shade_controller_impl
Change-Id: Ifcabfa17f859e2b3c135b4b410cf9b25884cef35