Remove mustHaveFlags argument from methods with a FlagChecker argument.
This is flagged because there is a very subtle behavior change, because "FlagChecker" was applied in places that "mustHaveFlags" was not. The only placed this was done where a flag was passed were FLAG_LIFETIME_EXTENDED_BY_DIRECT_REPLY and FLAG_AUTO_CANCEL, and the only understood application was that we didn't apply this constraint to children of the cancelled notification.  The CL recreates this behavior by checking the flag in advance where possible. 

There is one place where this is not done: when we post a CancelNotificationRunnable on a 200ms delay.  We need that FlagChecker to enforce the flag *after* the 200ms delay, so the FlagChecker does that, and that means that cancelled children will require that flag to be cancelled by interaction.  However, group summaries shouldn't ever be participating in direct reply, so this behavior change should not have any practical implications.

Bug: 440125528
Flag: android.app.nm_remove_must_have_flags
Test: atest NotificationManagerServiceTest
Change-Id: I89446d8fb4c390dec802ae6e455206049f55ec19