Use common ID for both variants of the empty shade
We're introducing a new empty shade design that includes an icon. We
need a way to uniquely identify this view in tests, regardless of
whether it's the new or the old design. We previously used
"no_notifications", but that doesn't exist in the new design, so we'll
use this new ID (see other CL in topic).

Flag: EXEMPT PURE_REFACTOR
Test: presubmit
Bug: 388472403
Change-Id: I3f48941c1ae93cffbcf75b451b1809dae7297e34