Skip desk deactivation on overview hide if display can't be found.
A crash occurs when disconnecting a display that has overview open due to an attempt to find a deactivate a desk that is no longer present on the display due to disconnect transition. To avoid this occurring, skip the deactivation altogether if the Display cannot be found, as this would mean that a disconnect has occurred.

Bug: 470878627
Test: Manual
Flag: com.android.window.flags.enable_display_disconnect_interaction
Change-Id: I1da21398dada1c2dff6a976c7b382facb9c511ad