Log whenever sysfs node monitoring did not start
We are seeing flakes of tests that check the battery functionality. One
observation is that in the successful cases, sysfs node monitoring
successfully started, but in the failing case, there is no such log.

In this CL, add a log whenever the input device for the requested id is
not found. In those cases, sysfs node monitoring does not start, which
could explain the flakiness that we are seeing.

Bug: 470344725
Test: none
Flag: EXEMPT DEBUG
Change-Id: Id929738f087042d78d0d6fe7be0b416624a10a3a