Always update stable bounds when drag start
Previously, the stable bounds were cached and updated only under
specific conditions: at the start of a drag if the display rotation had
changed, or at the end of a drag if the window had been moved to a
different display.

This logic was insufficient as it missed other critical display
configuration changes that can affect stable bounds, such as a change
in density, which could lead to incorrect resizing operations.

This change simplifies the logic by ensuring stable bounds are always
refetched at the beginning of every drag operation.

Flag: EXEMPT BUGFIX
Bug: 443009924
Test: atest MultiDisplayVeiledResizeTaskPositionerTest
Change-Id: If53ecdc1d44f8a9d1b964e332b6e2b8658c670a8