cleanup(mag always on): remove the flag class and fix the config/settings check
Since there is no need to remote control the always on feature flag, we
remove the flag check so now it only depend on the config value, which
is true by default and overridden to false on wear devices.

Besides, in AccessibilityManagerService we mis-treated the config value to
be the settings preference state. Therefore we fix the check, now the
settings preference is default on, and the wear can still disable the
functionalility by the config guard.

Also declare the always on settings preference default in
MagnificationConstants so that we can use that cross the platform and
settings package for consistency.

Bug: 472478320
Flag: EXEMPT BUGFIX
Test: manually check
      atest AccessibilityManagerServiceTest
Change-Id: I6400623577d06525e9815834a8e312500a0c2008