Be more accurate about when a HUN is displayed within the stack
During transitions into and out of shade or QS, we set the Z translation of a HUN based on the expansion fraction, but before this change we were only checking row.isAboveShelf(), which is true for pinned HUNs even when no transition is occurring (while also being true for all HUNs appearing when the stack is not present, such as over QS or when the shade is closed).

Setting Z based on expansion fraction (due to assuming the HUN was outside the stack) caused pinned HUNs to get a Z value of 0, which caused them to be below all the other notifications when scrolling. In the case of a pinned HUN appearing inside the stack, it should be more correct to use the "HUN is displayed within the stack" logic to determine the correct Z translation. This change checks the expansion fraction of the shade and QS in addition to isAboveShelf() to confirm that we're actually in a transition state to/from, or settled on, QS or Gone.

Bug: 470134431
Bug: 454601197
Test: StackScrollAlgorithmTest; manual by verifying pinned and non-pinned HUN behavior inside the stack, over QS, and over Gone
Flag: com.android.systemui.scene_container
Change-Id: I87c57226974a461ba2294e21b1b0d5371db29a25